+++ common/src/main/java/immersive_melodies/mixin/MobEntityMixin.java
@@ -7,9 +7,14 @@ import net.minecraft.entity.EntityType;
 import net.minecraft.entity.ItemEntity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.mob.MobEntity;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.Hand;
+import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.Vec3i;
 import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -25,8 +30,26 @@ public abstract class MobEntityMixin extends LivingEntity {
     @Shadow
     protected abstract void loot(ItemEntity item);
 
+    @Shadow
+    @Final
+    private DefaultedList<ItemStack> handItems;
+
     protected MobEntityMixin(EntityType<? extends LivingEntity> entityType, World world) {
         super(entityType, world);
+    }
+
+    @Inject(method = "interact(Lnet/minecraft/entity/player/PlayerEntity;Lnet/minecraft/util/Hand;)Lnet/minecraft/util/ActionResult;", at = @At("HEAD"), cancellable = true)
+    private void immersiveMelodies$injectInteract(PlayerEntity player, Hand hand, CallbackInfoReturnable<ActionResult> cir) {
+        for (ItemStack handItem : this.handItems) {
+            if (handItem.getItem() instanceof InstrumentItem) {
+                ItemEntity itemEntity = dropStack(handItem.copyAndEmpty());
+                if (itemEntity != null) {
+                    itemEntity.setThrower(getUuid());
+                }
+                cir.setReturnValue(ActionResult.CONSUME);
+                break;
+            }
+        }
     }
 
     @Inject(method = "tickMovement()V", at = @At("TAIL"))
@@ -34,7 +57,8 @@ public abstract class MobEntityMixin extends LivingEntity {
         if (Config.getInstance().forceMobsToPickUp && EntityEquiper.canPickUp(this) && !this.getWorld().isClient && this.isAlive() && !this.dead) {
             Vec3i vec3i = this.getItemPickUpRangeExpander();
             for (ItemEntity itementity : this.getWorld().getNonSpectatingEntities(ItemEntity.class, this.getBoundingBox().expand(vec3i.getX(), vec3i.getY(), vec3i.getZ()))) {
-                if (!itementity.isRemoved() && !itementity.getStack().isEmpty() && itementity.getStack().getItem() instanceof InstrumentItem) {
+                if ((itementity.getOwner() == null || !itementity.getOwner().getUuid().equals(getUuid())) && !itementity.isRemoved() && !itementity.getStack().isEmpty() && itementity.getStack().getItem() instanceof InstrumentItem) {
+                    System.out.println(itementity.getOwner());
                     this.loot(itementity);
                 }
             }