+++ changelog.md
@@ -1,6 +1,7 @@
 # 0.0.3
 
 * Only relevant entities pick up instruments
+* Right-click on an entity will cause them to drop their instrument
 
 # 0.0.2